BiocParallel-package {BiocParallel}	R Documentation

Bioconductor facilities for parallel evaluation

Description

This package provides modified versions and novel implementation of functions for parallel evaluation, tailored to use with Bioconductor objects.

Details

This package uses code from the parallel package,

Author(s)

Author: Bioconductor Package Maintainer [cre], Martin Morgan [aut], Valerie Obenchain [aut], Michel Lang [aut], Ryan Thompson [aut], Nitesh Turaga [aut]

Maintainer: Bioconductor Package Maintainer <maintainer@bioconductor.org>

[Package BiocParallel version 1.18.1 Index]