SerialParam-class {BiocParallel}	R Documentation

Enable serial evaluation

Description

This class is used to parameterize serial evaluation, primarily to facilitate easy transition from parallel to serial code.

Usage


SerialParam(stop.on.error = TRUE, log = FALSE,
    threshold = "INFO", logdir = NA_character_, progressbar = FALSE)

Arguments

stop.on.error
A logical determining behavior on error; see SnowParam.

log
logical(1) Enable logging; see SnowParam.

threshold
character(1) Logging threshold; see SnowParam.

logdir
character(1) Log files directory. When not provided, log messages are returned to stdout.

progressbar
logical(1) Enable progress bar (based on plyr:::progress_text).

Constructor

SerialParam():

Return an object to be used for serial evaluation of otherwise parallel functions such as bplapply, bpvec.

Methods

The following generics are implemented and perform as documented on the corresponding help page (e.g., ?bpworkers): bpworkers. bpisup, bpstart, bpstop, are implemented, but do not have any side-effects.

Author(s)

Martin Morgan mailto:mtmorgan@fhcrc.org

See Also

getClass("BiocParallelParam") for additional parameter classes.

register for registering parameter classes for use in parallel evaluation.

Examples


p <- SerialParam()
simplify2array(bplapply(1:10, sqrt, BPPARAM=p))
bpvec(1:10, sqrt, BPPARAM=p)

## Not run: 
register(SerialParam(), default=TRUE)

## End(Not run)

[Package BiocParallel version 1.18.1 Index]