BiocParallel-package {BiocParallel}	R Documentation

Bioconductor facilities for parallel evaluation

Description

This package provides modified versions and novel implementation of functions for parallel evaluation, tailored to use with Bioconductor objects.

Details

This package uses code from the parallel package,

Author(s)

See packageDescription("BiocParallel").

[Package BiocParallel version 1.28.3 Index]